+++ CHANGES.rst
@@ -7,6 +7,7 @@ next release
 ------------
 - Update release tooling
 - Upgrade to materialize-css 1.0.0
+- Adjust URL hash to support browsing state/history
 
 2019-11-06 0.7.0
 ----------------